Efficiency and Adaptation Fixes
Basically, take care of every warning and use macros to calculate some
of the timing info (in case, for example, we switch the CPU frequency).

Contains:

Revert "Efficiency and Adaptation Fixes"

This reverts commit ab1e049663e7903c286c238bc90c3768860c8026.

There seems to be some small bugs. Since the main issue was found and
fixed, let's just stick to the older code for the most part.

Add an in-betweener buffer

Just in case the MQTT sending takes longer than one full buffer time.
Shouldn't be able to happen, but just in case...

Better Cleanup Fix Set

fixup! Efficiency and Adaptation Fixes