fixup! Disable broker timeouts on DHCP
DHCP takes much longer to configure the connection than manual IP
config, and that's long enough that the MQTT timeout will probably get
triggered.

Can't use the CLI to do this because it isn't initialized yet when the
Network is configured).

Also ensure that the MQTT service setup doesn't override this value