feat: Shell Handling Library for Line Editing (#7)
Adds vendored `microrl-remaster` library to handle line editing before passing the commands on to `Commander`.

PlatformIO kept exploding when trying to build the `microrl-remastered`
library because of the folder structure. Resolved by vendoring v2.5.0
and including our `_user_config.h` in the source folder, finally making
the compile process happy.

Moves the `adc dump` ending check into `cli_loop()` so that both streams
can be monitored. Also gets away from checking `cli->getInputPort()`
which is now a NULL to work with `microrl`. Since that created too much
potential for an extra character in the buffer to stop the streaming prematurely,
we instead just wait until a command is executed.

Includes: 
* shell-lib: use StreamUtils for both StringStream and Tee
* chore: add NOTICES for vendored microrl library